+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 # Changelog
 
+## 0.30.5
+
+- bug fixes
+  - regression breaking multitouch widgets
+
 ## 0.30.4
 
 - bug fixes
   - regression breaking touch traversing gestures
   - clone widgets becomming unselectable for edition
   - `preArgs` property still being coerced to an array
-  
+
 ## 0.30.3
 
 - bug fixes